+++ src/main/java/com/researchspace/figshare/impl/FigshareTemplate.java
@@ -26,6 +26,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.client.BufferingClientHttpRequestFactory;
+import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.http.converter.ByteArrayHttpMessageConverter;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
@@ -79,9 +81,15 @@ public final class FigshareTemplate implements Figshare {
 	}
 
 	private void init() {
-		this.restTemplate = new RestTemplate();
+		this.restTemplate = new RestTemplate(
+				new BufferingClientHttpRequestFactory(new SimpleClientHttpRequestFactory()));
 		configureRestTemplate();
-		this.fileOps = new FileOperationsImpl(restTemplate, accessToken);
+		// File operations get a plain (non-buffered) RestTemplate to avoid
+		// duplicating large upload payloads in memory. DefaultResponseErrorHandler
+		// (the default) is used so errors throw proper Spring exceptions without
+		// consuming/closing the response stream.
+		RestTemplate fileRestTemplate = new RestTemplate();
+		this.fileOps = new FileOperationsImpl(fileRestTemplate, accessToken);
 		this.utils = new FigshareUtils();
 	}
 
@@ -98,7 +106,7 @@ public final class FigshareTemplate implements Figshare {
 	public Location createArticle(ArticlePost article) {
 		String url = utils.createPath("/account/articles");
 		String json = marshalObject(article);
-		HttpEntity<String> entity = utils.createHttpEntity(json, accessToken);
+		HttpEntity<String> entity = utils.createJsonHttpEntity(json, accessToken);
 		ResponseEntity<String> resp = getRestTemplate().postForEntity(url, entity, String.class);
 		log.debug(resp.toString());
 		// can't convert directly as content type of returned is text/html?
@@ -110,7 +118,7 @@ public final class FigshareTemplate implements Figshare {
 	public Location createFile(Long articleId, File file) {
 		String url = utils.createPath("/account/articles/{articleId}/files");
 		ObjectNode node = createFileJson(file);
-		HttpEntity<String> entity = utils.createHttpEntity(marshalObject(node), accessToken);
+		HttpEntity<String> entity = utils.createJsonHttpEntity(marshalObject(node), accessToken);
 		ResponseEntity<String> resp = getRestTemplate().postForEntity(url, entity, String.class, articleId);
 		log.debug(resp.toString());
 		// can't convert directly as content type of returned is text/html?